clean_xspf_nops.py
This utility removes the superfluous "vlc:node" entries in XSPF playlists,
which get created when saving over an old/existing playlist and "tree mode"
is enabled for the playlist display. The problem is that, while it's useful
to be able to see the name of the playlist that a bunch of files come from
when playing back in the UI, if you save over an existing playlist (after
adding 1 or more extra tracks), VLC will insert extra "folders" in the playlist,
denoted using "vlc:node" elements and corresponding to the playlist that you've
saved over. However, when reloading this file, these extra folders serve no
real purpose (other than further deep-nesting and hiding the actual tracks)
and cannot be removed from the UI without removing the child items.